Copy a custom .repo file first and refer to that copy from now on
Since one can have their own .repo file for offline migration anywhere,
there might be a case that it resides in /etc/yum.repos.d which
conflicts with deleting all .repo files (if no 'preserve' option was
provided). Copy it to a safe location and use the copy as the reference.